requirement.txt
altgraph==0.16.1
astroid==2.0.4
beautifulsoup4==4.6.0
bokeh==1.0.2
bottle==0.12.15
certifi==2018.10.15
cffi==1.11.5
chardet==3.0.4
Click==7.0
colorama==0.4.0
confusable-homoglyphs==3.2.0
dj-database-url==0.5.0
Django==2.1.4
django-classy-tags==0.8.0
django-cms==3.5.3
django-contact-us==0.4.1
django-dirtyfields==1.3.1
django-filer==1.4.1
django-formtools==2.1
django-js-asset==1.1.0
django-modelcluster==4.2
django-mptt==0.9.1
django-parsley==0.7
django-plugins==0.3.0
django-polymorphic==2.0.3
django-registration==3.0
django-sekizai==0.10.0
django-tables2==2.0.3
django-taggit==0.23.0
django-tinymce==2.8.0
django-treebeard==4.3
djangocms-admin-style==1.2.9
djangocms-attributes-field==1.0.0
djangocms-column==1.9.0
djangocms-file==2.2.0
djangocms-googlemap==1.2.0
djangocms-installer==1.0.2
djangocms-link==2.3.1
djangocms-picture==2.1.3
djangocms-snippet==2.1.0
djangocms-style==2.1.0
djangocms-text-ckeditor==3.7.0
djangocms-video==2.1.1
djangorestframework==3.9.0
draftjs-exporter==2.1.2
easy-thumbnails==2.5
future==0.17.1
gmail==0.6.3
greenlet==0.4.15
html2text==2018.1.9
html5lib==1.0.1
idna==2.7
isort==4.3.4
Jinja2==2.10
lazy-object-proxy==1.3.1
macholib==1.11
MarkupSafe==1.1.0
mccabe==0.6.1
mysql-connector==2.1.6
numpy==1.15.4
packaging==18.0
pefile==2018.8.8
Pillow==5.3.0
pycparser==2.19
Pygments==2.3.0
PyInstaller==3.4
pylint==2.1.1
pyparsing==2.3.0
pyqode.core==2.11.0
pyqode.designer==1.2
pyqode.qt==2.10.0
PyQRCode==1.2.1
PyQt5==5.11.3
PyQt5-sip==4.19.13
pyqt5-tools==5.11.2.1.3
PyQt5Designer==5.10.1
python-dateutil==2.7.5
python-dotenv==0.9.1
pytz==2018.5
pywin32-ctypes==0.2.0
PyYAML==3.13
reportlab==3.5.19
requests==2.20.0
six==1.11.0
sockets==1.0.0
South==1.0.2
subprocess.run==0.0.8
tornado==5.1.1
tzlocal==1.5.1
Unidecode==1.0.22
urllib3==1.24
virtualenv==16.0.0
wagtail==2.2.2
webencodings==0.5.1
Willow==1.1
wrapt==1.10.11
